We mostly copy the "expert" layout from the React frontend at https://github.com/hackworthltd/primer-app very closely.
This code is _much_ more concise than the equivalent TypeScript/React/Tailwind code, even if the lack of support for non-expert levels makes a precise comparison difficult. This is down to the power of Haskell, the hindsight afforded by a rewrite, and Tailwind's lack of compositionality along certain axes (it was impossible to really factor out using the same color as a part of multiple attributes, since Tailwind doesn't allow concatenating strings to form class names). In reverse-engineering that mess of code and teasing out its structure, we uncover several oddities which we don't replicate here:
- We set some text colors which are completely overridden, such as `text-grey-tertiary` for most nodes.
- We used `text-red-primary` instead of `text-red-tertiary` for type-level holes.
- We set the background twice instead of ever setting the border for `KType` and `KFun`, in `flavorLabelClasses`.

Other differences:
- We show the full word "match" on pattern match nodes, instead of just "m". This was always wrong but awkward to fix with the old architecture.
- We scale kind nodes down slightly rather than telling the layout engine to treat them as a different size to other syntax nodes. This is slightly easier to implement, and arguably looks better.
- We use colours for kinds consistent with type/term analogues, e.g. `KHole` is now red. Again, this is somewhat implementation driven (the new code, being much less ad-hoc, generally makes consistency the easier option), but is also probably an improvement.

We also now layout based on actual node widths and heights. This could have probably been made a separate commit, but there were advantages to developing both features alongside each other. The styled nodes would have been difficult to view without the improved layout, and the requirements for getting the information needed for the layout engine had some impact on how we approached styling.

We avoid changing the tree structure, since we're already doing a lot in one commit. Therefore better rendering of bindings and patterns will wait for a follow-up.