Draw bindings separately from parent syntax nodes
Previously lambdas and foralls had the text baked in, in a way that was frequently ugly, and let/let-type/let-rec simply didn't show the bindings at all (there are no lets in the current example, so this was overlooked).

We'd ideally show these off to the right, as with the React frontend. But rather than resorting to the hacks we used there, it's probably best to wait until we have a layout engine which understands a notion of horizontal children.

Note this applies for now only to bindings in lambdas, foralls and lets, but _not_ patterns, which will be handled separately.