Render patterns in boxes
Previously, patterns were rendered in a similar way to our early prototype Vonnegut, and were thus difficult to read. This updates them to be more like our React frontend. The only difference is that, as with variable bindings, we render them as ordinary child nodes, rather than off to the right. This is again down to simplicity rather than preference.

It turns out the layout algorithm we're using is not as smart as the Tidy algorithm (used in the React frontend) when it comes to packing siblings with different heights, but it will do for now.